# app/api_spec.yml
openapi: '3.0.3'
info:
  title: FHIR Genomics Operations
  version: '0.1'
  description: |-
    This site is a reference implementation of FHIR Genomics Operations.

    FHIR Genomics operations are based on the premise that genomic data, in FHIR format and/or some other format (e.g. <a href="https://samtools.github.io/hts-specs/VCFv4.3.pdf">VCF format</a>), are stored in a repository, either in or alongside an EHR, possibly along with phenotype annotations. The FHIR Genomics operations essentially 'wrap' the repository, presenting a uniform interface to applications, regardless of internal repository data structures.

    Click <a href="http://build.fhir.org/ig/HL7/genomics-reporting/operations.html">here</a> to Learn more about FHIR Genomics Operations

    All patient data used in this implementation is either synthetic or publicly available and anonymized. All data and code used in this implementation is open-sourced. Click <a href="https://github.com/FHIR/genomics-operations">here</a> to learn more about the reference implementation.

    Contact <a href="mailto:info@elimu.io">info@elimu.io</a> for comments on this Reference Implementation.
servers:
  - url: https://fhir-gen-ops.herokuapp.com
paths:
...
  /utilities/get-feature-coordinates:
    get:
      description: 'This utility returns genomic feature coordinates and other annotations. All data are from <a href="https://www.ncbi.nlm.nih.gov/genome/guide/human/">NCBI Human Genome Resources</a>. For chromosomes, build 37 and build 38 reference sequences are returned. For genes, genomic coordinates are returned, along with a list of transcripts. MANE transcript is flagged. For transcripts, genomic coordinates are returned, along with the gene name and composite exons, along with exon coordinates. For proteins, the corresponding transcript is returned.'
      summary: "Get Feature Coordinates"
      operationId: "app.utilities_endpoints.get_feature_coordinates"
      tags:
        - "Operations Utilities (not part of balloted HL7 Operations)"
      responses:
        '200':
          description: ""
          content:
            application/json:
              schema:
                type: object
      parameters:
        - name: chromosome
          in: query
          description: Valid entries include chr1..chr22, chrX, chrY, chrM.
          schema:
            type: string
            pattern: '^\s*\b(chr|Chr|CHR)([1-9]|1[0-9]|2[0-2]|[XxYyMm])\s*$'
          example: "chr1"
        - name: gene
          in: query
          description: Valid entries include an HGNC gene symbol or an HGNC gene ID (optionally prefixed with ‘HGNC:’).
          schema:
            type: string
          example: "HGNC:5"
        - name: transcript
          in: query
          description: Valid entries are NCBI transcript reference sequences (starting with ‘NM_’). (Version numbers are ignored - the latest version is always returned).
          schema:
            type: string
            pattern: '^\s*[Nn][Mm]_\d{4,10}(\.)?(\d{1,2})?\s*$'
          example: "NM_000014.6"
        - name: protein
          in: query
          description: Valid entries are NCBI protein reference sequences (starting with ‘NP_’). (Version numbers are ignored - the latest version is always returned).
          schema:
            type: string
            pattern: '^\s*[Nn][Pp]_\d{4,10}(\.)?(\d{1,2})?\s*$'
          example: "NP_000005.3"
  /utilities/find-the-gene:
    get:
      description: |-
        This utility returns all genes that intersect with a provided genomic region.
      summary: "Find The Gene"
      operationId: "app.utilities_endpoints.find_the_gene"
      tags:
        - "Operations Utilities (not part of balloted HL7 Operations)"
      responses:
        '200':
          description: "Returns all genes that intersect with a provided genomic region."
          content:
            application/json:
              schema:
                type: object
      parameters:
        - name: range
          in: query
          required: true
          description: Genomic region to be searched for genes. Must be in zero-based RefSeq:Integer-range format.
          schema:
            type: string
            pattern: '^\s*[Nn][Cc]_\d{4,10}(\.)(\d{1,2}):\d{1,10}-\d{1,10}\s*$'
            example: "NC_000001.11:11794399-11794400"
  /utilities/normalize-variant:
    get:
      description: >
        This utility uses <a href="https://api.ncbi.nlm.nih.gov/variation/v0/">NCBI Variation Services</a> to convert an arbitrary HGVS or SPDI-formatted variant into a canonical build 37 and build 38 HGVS and SPDI.
      summary: "Normalize variant"
      operationId: "app.utilities_endpoints.normalize_variant"
      tags:
        - "Operations Utilities (not part of balloted HL7 Operations)"
      responses:
        "200":
          description: "Returns a canonical build 37 and build 38 HGVS and SPDI."
          content:
            application/json:
              schema:
                type: object
      parameters:
        - name: variant
          in: query
          required: true
          description: "Variant must be in HGVS or SPDI format."
          schema:
            type: string
            example: "NM_001127510.3:c.145A>T"
  /utilities/normalize-hla:
    get:
      description: >
        This utility uses the <a href="https://pypi.org/project/py-ard/">python py-ard library</a> to convert an arbitrary HLA allele into various canonical forms.
      summary: "Normalize HLA"
      operationId: "app.utilities_endpoints.normalize_hla"
      tags:
        - "Operations Utilities (not part of balloted HL7 Operations)"
      responses:
        "200":
          description: "Returns a normalized HLA ARD allele."
          content:
            application/json:
              schema:
                type: object
      parameters:
        - name: allele
          in: query
          required: true
          description: "Allele."
          schema:
            type: string
            example: "B14"
  /utilities/translate-terminology:
    get:
      description: |-
        This utility demonstrates terminology translations SNOMED to MedGen and DiseaseOntology; ICD10 to MedGen and DiseaseOntology; RxNorm to NCIt.
      summary: "Translate from one code system to another"
      operationId: "app.utilities_endpoints.translate_terminology"
      tags:
        - "Operations Utilities (not part of balloted HL7 Operations)"
      responses:
        '200':
          description: "Returns matching codes in target code system(s)."
          content:
            application/json:
              schema:
                type: object
      parameters:
        - name: codeSystem
          in: query
          required: true
          description: Code system for code to be translated.
          schema:
            type: string
            enum:
              - "http://www.nlm.nih.gov/research/umls/rxnorm"
              - "http://snomed.info/sct"
              - "http://hl7.org/fhir/sid/icd-10"
            example: "http://snomed.info/sct"
        - name: code
          in: query
          schema:
            type: string
            example: '126949007'
          required: true
          description: The code to be translated.
  /utilities/parse-genomic-study:
    get:
      description: >
        Parse a FHIR R5 GenomicStudy JSON file into structured analysis data.
      summary: "Parse Genomic Study"
      operationId: "app.utilities_endpoints.parse_genomic_study"
      tags:
        - "Operations Utilities (not part of balloted HL7 Operations)"
      responses:
        '200':
          description: "Returns extracted analyses from the GenomicStudy resource."
          content:
            application/json:
              schema:
                type: object
      parameters:
        - name: file
          in: query
          required: true
          description: "Path to the GenomicStudy JSON file on the server."
          schema:
            type: string
            example: "/data/genomic_study.json"
...